ensure mqtt doesn't get linted by wasm lint config
