+++ crates/mqtt/Cargo.toml
@@ -10,8 +10,10 @@ homepage.workspace = true
 
 [dependencies]
 egui_plot.workspace = true
-rumqttc.workspace = true
 log.workspace = true
+
+[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
+rumqttc.workspace = true
 
 [lints]
 workspace = true
