+++ crates/mqtt/src/lib.rs
@@ -1,18 +1,17 @@
 use egui_plot::PlotPoint;
-use rumqttc::{Client, Event, MqttOptions, Packet, QoS};
-use std::{
-    sync::{
-        atomic::{AtomicBool, Ordering},
-        mpsc, Arc,
-    },
-    time::Duration,
-};
-
+#[cfg(not(target_arch = "wasm32"))]
 pub(crate) mod broker_validator;
-pub mod data_receiver;
-pub mod mqtt_cfg_window;
+#[cfg(not(target_arch = "wasm32"))]
+pub(crate) mod mqtt_listener;
+#[cfg(not(target_arch = "wasm32"))]
 pub(crate) mod topic_discoverer;
+#[cfg(not(target_arch = "wasm32"))]
 pub(crate) mod util;
+
+#[cfg(not(target_arch = "wasm32"))]
+pub mod data_receiver;
+#[cfg(not(target_arch = "wasm32"))]
+pub mod mqtt_cfg_window;
 
 /// Accumulated plot points from an MQTT topic
 #[derive(Debug)]
@@ -26,63 +25,4 @@ pub struct MqttPoints {
 pub struct MqttPoint {
     pub topic: String,
     pub point: PlotPoint,
-}
-
-pub fn mqtt_listener(
-    tx: &mpsc::Sender<MqttPoint>,
-    broker: String,
-    topics: Vec<String>,
-    stop_flag: &Arc<AtomicBool>,
-) {
-    let timestamp_id = std::time::SystemTime::now()
-        .duration_since(std::time::SystemTime::UNIX_EPOCH)
-        .expect("Time went backwards")
-        .as_millis();
-    let mut mqttoptions = MqttOptions::new(format!("plotinator3000-{timestamp_id}"), broker, 1883);
-    mqttoptions.set_keep_alive(Duration::from_secs(5));
-
-    let (client, mut connection) = Client::new(mqttoptions, 10);
-    for t in topics {
-        if let Err(e) = client.subscribe(t, QoS::AtMostOnce) {
-            log::error!("Subscribe error: {e}");
-        }
-    }
-
-    // Iterate to poll the eventloop for connection progress
-    for notification in connection.iter() {
-        if stop_flag.load(Ordering::Relaxed) {
-            log::info!("Stopping!");
-            break;
-        }
-        match notification {
-            Ok(event) => {
-                if let Event::Incoming(Packet::Publish(publish)) = event {
-                    let topic = publish.topic;
-                    let payload = String::from_utf8_lossy(&publish.payload);
-                    log::info!("Received on topic={topic}, payload={payload}");
-                    match payload.parse::<f64>() {
-                        Ok(num) => {
-                            let now = std::time::SystemTime::now()
-                                .duration_since(std::time::UNIX_EPOCH)
-                                .expect("Time went backwards")
-                                .as_nanos() as f64;
-                            log::info!("now={now}");
-                            let point = PlotPoint::new(now, num);
-                            let mqtt_data = MqttPoint { topic, point };
-                            if let Err(e) = tx.send(mqtt_data) {
-                                log::error!("Send err={e}");
-                            }
-                        }
-                        Err(e) => log::error!("Payload parse error: {e}"),
-                    }
-                }
-            }
-
-            Err(e) => log::error!("{e}"),
-        }
-    }
-    if let Err(e) = client.disconnect() {
-        log::error!("{e}");
-        debug_assert!(false, "{e}");
-    }
 }
